+++ ui/src/loudml/components/ParametersPanel.js
@@ -5,45 +5,55 @@ const ParametersPanel = ({
     model,
     onInputChange,
 }) => {
-  return (
-    <div className="panel panel-solid">
-        <div className="panel-heading">
-            <h2 className="panel-title"></h2>
-        </div>
-        <div className="panel-body">
-            <div className="col-xs-6">
-                <div className="form-group">
+    function handleSeasonality(e) {
+        const {name, id, checked} = e.target
+    
+        onInputChange({
+            target: {
+                name,
+                type: 'custom',
+                value : {
+                    ...model.seasonality,
+                    [id]: checked,
+                }
+            }
+        })
+    }
+    
+    return (
+        <div className="panel panel-solid">
+            <div className="panel-heading">
+                <h2 className="panel-title"></h2>
+            </div>
+            <div className="panel-body">
+                <div className="form-group col-xs-12 col-sm-6">
                     <label htmlFor="max_evals">Max iterations</label>
                     <input
                         type="number"
                         name="max_evals"
-                        className="form-control input-sm form-malachite"
+                        className="form-control input-md form-malachite"
                         value={model.max_evals}
                         onChange={onInputChange}
                         placeholder="ex: 100"
                     />
                 </div>
-            </div>
-            <div className="col-xs-6">
-                <div className="form-group">
+                <div className="form-group col-xs-12 col-sm-6">
                     <label htmlFor="bucket_interval">Bucket interval</label>
                     <input
                         type="text"
                         name="bucket_interval"
-                        className="form-control input-sm form-malachite"
+                        className="form-control input-md form-malachite"
                         value={model.bucket_interval}
                         onChange={onInputChange}
                         placeholder="ex: 30s, 20m, 1h, 1d, ..."
                     />
                 </div>
-            </div>
-            <div className="col-xs-6">
-                <div className="form-group">
+                <div className="form-group col-xs-12 col-sm-6">
                     <label htmlFor="span">Span</label>
                     <input
                         type="number"
                         name="span"
-                        className="form-control input-sm form-malachite"
+                        className="form-control input-md form-malachite"
                         value={model.span}
                         onChange={onInputChange}
                         placeholder="ex: 5"
@@ -51,9 +61,33 @@ const ParametersPanel = ({
                         required={true}
                     />
                 </div>
+                <div className="form-group col-xs-12 col-sm-6">
+                    <label>Seasonality</label>
+                    <div className="form-control-static">
+                        <div className="col-xs-12 col-sm-6">
+                            <input
+                            type="checkbox"
+                            id="daytime"
+                            name="seasonality"
+                            checked={model.seasonality.daytime}
+                            onChange={handleSeasonality}
+                                />
+                            <label htmlFor="daytime">Daytime</label>
+                        </div>
+                        <div className="col-xs-12 col-sm-6">
+                            <input
+                                type="checkbox"
+                                id="weekday"
+                                name="seasonality"
+                                checked={model.seasonality.weekday}
+                                onChange={handleSeasonality}
+                                />
+                                <label htmlFor="weekday">Weekday</label>
+                        </div>
+                    </div>
+                </div>
             </div>
         </div>
-    </div>
     )
 }
